Add files via upload
We add the three main approaches implemented. The naive one, which iterates through dimensions, then utilizes a set of candidate amount of points to finde the minimal number of points on each dimension to get the desired persistence in the desired homology group. The perallel one, which builds on top of the naive one, but parallelizing the monte carlo simulation, reducing the required time to exclude a given number of points. And the suspension approach, which builds on top the the previous two, it relies on suspension theory to create a layered nsphere, using a (n-1)-sphere on each layer, it also perfom a parallelized montecarlo simulation.